requirements.txt
Flask~=2.1.2
GitPython~=3.1.27
RPi.GPIO~=0.7.1